+++ backend/src/post/post.controller.ts
@@ -7,42 +7,30 @@ import {
   Param,
   Delete,
   UploadedFile,
+  UseInterceptors,
 } from "@nestjs/common";
 import { PostService } from "./post.service";
 import { CreatePostDto } from "./dto/create-post.dto";
 import { UpdatePostDto } from "./dto/update-post.dto";
 import { ApiOkResponse, ApiOperation, ApiTags } from "@nestjs/swagger";
 import { PostModel } from "./post.model";
+import { FileInterceptor } from "@nestjs/platform-express";
 
 @ApiTags("Posts functional")
 @Controller("posts")
 export class PostController {
   constructor(private readonly postService: PostService) {}
 
-  @ApiOperation({
-    summary: "Create new post",
-  })
-  @ApiOkResponse({
-    status: 200,
-    type: PostModel,
-  })
+  @ApiOperation({ summary: "Create new post" })
+  @ApiOkResponse({ status: 200, type: PostModel })
   @Post()
-  create(
-    @Body()
-    createPostDto: CreatePostDto,
-    @UploadedFile()
-    image
-  ) {
+  @UseInterceptors(FileInterceptor("image"))
+  create(@Body() createPostDto: CreatePostDto, @UploadedFile() image) {
     return this.postService.create(createPostDto, image);
   }
 
-  @ApiOperation({
-    summary: "Find all posts",
-  })
-  @ApiOkResponse({
-    status: 200,
-    type: [PostModel],
-  })
+  @ApiOperation({ summary: "Find all posts" })
+  @ApiOkResponse({ status: 200, type: [PostModel] })
   @Get()
   getAllPosts() {
     return this.postService.findAll();
@@ -72,12 +60,9 @@ export class PostController {
   })
   @Patch("/:id")
   update(
-    @Param("id")
-    id: string,
-    @Body()
-    updatePostDto: UpdatePostDto,
-    @UploadedFile()
-    image
+    @Param("id") id: string,
+    @Body() updatePostDto: UpdatePostDto,
+    @UploadedFile() image
   ) {
     return this.postService.update(+id, updatePostDto, image);
   }
@@ -111,18 +96,10 @@ export class PostController {
     return this.postService.like(+id);
   }
 
-  @ApiOperation({
-    summary: "Unlike a post",
-  })
-  @ApiOkResponse({
-    status: 200,
-    type: PostModel,
-  })
+  @ApiOperation({ summary: "Unlike a post" })
+  @ApiOkResponse({ status: 200, type: PostModel })
   @Delete("/like/:id")
-  unlike(
-    @Param("id")
-    id: string
-  ) {
+  unlike(@Param("id") id: string) {
     return this.postService.unlike(+id);
   }
 }