+++ frontend/src/types/post/post.dto.ts
@@ -2,7 +2,7 @@ export interface PostDto {
     id: number;
     title: string;
     content: string;
-    image: string | string[];
+    image: string;
     likesCount: number;
     userId: number;
 }